Refactor create project page
